+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Unreleased
 
+- removed: Disable `accelerate` for bip44 wallets
+
 ## 2.5.3 (2024-01-24)
 
 - changed: Add insufficient funds check `subtractFee` style transactions